+++ resources/ts/interfaces/CardItemInterface.ts
@@ -1,5 +1,5 @@
 export default interface CardItemInterface {
     title: string
-    src: string
-    href: string
+    link: string
+    img: string
 }