+++ resources/ts/templates/cardItemTemplate.ts
@@ -2,9 +2,9 @@ import CardItemInterface from "../interfaces/CardItemInterface"
 
 export default (card: CardItemInterface): string => {
     return `
-        <a href="${card.href}" class="revival-random-video">
-            <img src="${card.src}" alt="${card.title}">
-            <span>${card.title}}</span>
+        <a href="${card.link}" class="revival-random-video" target="_blank">
+            <img src="${card.img}" alt="${card.title}">
+            <span>${card.title}</span>
         </a>
     `
 }